pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ffclients"
version = "0.1.0"
description = "AI Client Library for Far Finer LLC"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Antonio Quinonez", email = "antquinonez@farfiner.com"}
]

dependencies = [
    "mistralai",
    "anthropic",
    "openai",
    "azure-ai-inference",
    "azure-core",
    "google-auth",
    "duckdb",
    "requests",
    "pandas",
    "polars[rtcompat]",
    "python-dotenv",
    "nest_asyncio",
    "matplotlib",
    "openpyxl",
    "pyarrow",
    "litellm>=1.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
    "ruff",
    "invoke>=2.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "llamaparse: tests requiring LlamaParse API access",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "src.Clients*", "src.workbook*"]

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]
exclude = [
    "src/misc_code",
    "scripts",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument
    "B905",   # zip without explicit strict
    "B904",   # raise from err (common pattern in this codebase)
    "ARG001", # unused function argument (common in callbacks)
    "ARG002", # unused method argument
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "SIM", "E712", "F841", "F811", "SIM117", "C401"]
"src/Clients/*" = ["B904", "F841", "UP038"]
"tasks.py" = ["ARG001", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
